+++ PowerEditor/src/ScitillaComponent/AutoCompletion.h
@@ -80,7 +80,7 @@ public:
 
 	void insertMatchedChars(int character, const MatchedPairConf & matchedPairConf);
 	void update(int character);
-	void callTipClick(int direction);
+	void callTipClick(size_t direction);
 	void getCloseTag(char *closeTag, size_t closeTagLen, size_t caretPos, bool isHTML);
 
 private:
